requirements.txt
matplotlib==3.5.1
numpy==1.21.5
pandas==1.3.5
Pillow==8.4.0
pytorch_lightning==1.5.6
PyYAML==6.0
sentence_transformers==2.1.0
torch==1.10.1
torchvision==0.11.2
tqdm==4.62.3
transformers==4.14.1
